requirements.txt
Flask==1.1.2
gunicorn==20.0.4
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0